feat: added sharable query string (#39)
* feat: added query string for sharing
* feat: copy link from results modal
* feat: add copy chip
