+++ apps/web/src/components/EventCard.tsx
@@ -1,12 +1,29 @@
-import { VStack, Image, Box, Text, HStack, Button, useDisclosure } from '@chakra-ui/react';
+import {
+    VStack,
+    Image,
+    Box,
+    Text,
+    HStack,
+    Button,
+    useDisclosure,
+    useToast,
+} from '@chakra-ui/react';
 import React from 'react';
 import { DocumentData, QueryDocumentSnapshot } from 'firebase/firestore';
+import { useSearchParams } from 'react-router-dom';
 import { ResultsModal } from '../modal';
+import Toast from './Toast';
 
 const EventCard = ({ event }: EventCardProps) => {
     const { isOpen, onOpen, onClose } = useDisclosure();
-
+    const [, setSearchParams] = useSearchParams();
+    const toast = useToast();
     const { about, results, image, moreInfo, projectCount } = event.data();
+    const openResultModal = () => {
+        setSearchParams({ id: event.id });
+        onOpen();
+    };
+
     return (
         <>
             {isOpen && (
@@ -18,7 +35,47 @@ const EventCard = ({ event }: EventCardProps) => {
                 alignItems="flex-start"
                 borderRadius="10px"
             >
-                <Box backgroundColor="white" padding="30px" width="100%" borderTopRadius="10px">
+                <Box
+                    position="relative"
+                    backgroundColor="white"
+                    padding="30px"
+                    width="100%"
+                    borderTopRadius="10px"
+                >
+                    <Text
+                        display={results ? 'block' : 'none'}
+                        position="absolute"
+                        cursor="pointer"
+                        paddingInline="8px"
+                        right="10px"
+                        top="10px"
+                        fontFamily="Clash Display"
+                        paddingBlock="4px"
+                        borderRadius="10px"
+                        textColor="white"
+                        backgroundColor="rgba(0,0,0,.5)"
+                        _hover={{
+                            textColor: 'black',
+                            boxShadow: '0px 8px 16px rgba(255, 255, 255, 0.15)',
+                            backgroundColor: '#DBF72C',
+                        }}
+                        onClick={() =>
+                            navigator.clipboard
+                                .writeText(`${window.location.href}?id=${event.id}`)
+                                .then(() => {
+                                    toast({
+                                        title: '✅Copied to clipboard!',
+                                        status: 'success',
+                                        position: 'bottom',
+                                        render: ({ title, status }) => (
+                                            <Toast title={title} status={status} />
+                                        ),
+                                    });
+                                })
+                        }
+                    >
+                        Copy Link
+                    </Text>
                     <Image height="120px" src={image} objectFit="cover" />
                 </Box>
                 <VStack
@@ -69,7 +126,7 @@ const EventCard = ({ event }: EventCardProps) => {
                                 boxShadow: '0px 8px 16px rgba(219, 247, 44, 0.15)',
                                 backdropFilter: 'blur(25px)',
                             }}
-                            onClick={onOpen}
+                            onClick={openResultModal}
                         >
                             View Projects
                         </Button>
