+++ apps/web/src/modal/components/index.ts
@@ -1,0 +1,1 @@
+export { Items } from './ResulItems';
